EDIT CIERRE MENSUAL Y PLAN MENSUAL, path_logo

+++ obras/obras_proy_delete.php
@@ -5,6 +5,8 @@ $where_c_coste=" id_c_coste={$_SESSION['id_c_coste']} " ;
  //echo "El filtro es:{$_GET["filtro"]}"; 
 
 $id_obra=$_GET["id_obra"]  ;
+$solo_udo_vacias= isset($_GET["solo_udo_vacias"]) ? $_GET["solo_udo_vacias"] : '' ;
+
 
 require_once("../../conexion.php");
 require_once("../include/funciones.php");
@@ -18,20 +20,35 @@ if ($rs=Drow( 'OBRAS', "ID_OBRA=$id_obra AND $where_c_coste"))
    $id_obra = $rs["ID_OBRA"] ; // confirmacion de SEGURIDAD
    $id_prod_estudio_costes = $rs["id_prod_estudio_costes"] ;
    $id_subobra_auto = $rs["id_subobra_auto"] ;
- 
+  
    $sql= "DELETE FROM `PRODUCCIONES_DETALLE`  WHERE ID_PRODUCCION=$id_prod_estudio_costes "  ;     // vaciamos la id_produccion_estudio_coste solamente, las otras RV , por seguridad, deben de hacerlo manualmente    
    $Conn->query($sql) ;
   
-    
-   $sql= "DELETE FROM `Udos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
-   $Conn->query($sql) ;
+//   $solo_udo_vacias_where= $solo_udo_vacias ? " AND MED_PROYECTO=0 " : ""  ;   
    
-   $sql= "DELETE FROM `Capitulos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
-   $Conn->query($sql) ;
+   if($solo_udo_vacias)
+   {    
+       $sql= "DELETE FROM `Udos`  WHERE ID_OBRA=$id_obra  AND MED_PROYECTO=0 "  ;     // ejecuto DELETE
+       $Conn->query($sql) ;
 
-   $sql= "DELETE FROM `SubObras`  WHERE ID_OBRA=$id_obra AND ID_SUBOBRA<>'$id_subobra_auto' "  ;     // ejecuto DELETE
-   $Conn->query($sql) ;
+       $sql= "DELETE FROM `Capitulos`  WHERE ID_OBRA=$id_obra AND (ID_CAPITULO NOT IN (SELECT DISTINCT ID_CAPITULO FROM Udos WHERE ID_OBRA=$id_obra )) "  ;     // ejecuto DELETE
+       logs( "SQL CAPITULOS:" . $sql ) ;
+       logs( "SQL CAPITULOS:" . $Conn->query($sql) ) ;
 
+       $sql= "DELETE FROM `SubObras`  WHERE ID_OBRA=$id_obra AND ID_SUBOBRA<>'$id_subobra_auto' AND ID_SUBOBRA NOT IN (SELECT DISTINCT  ID_SUBOBRA FROM Udos WHERE ID_OBRA=$id_obra)"  ;     // ejecuto DELETE
+       $Conn->query($sql) ;
+
+    }else
+    {
+       $sql= "DELETE FROM `Udos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
+       $Conn->query($sql) ;
+
+       $sql= "DELETE FROM `Capitulos`  WHERE ID_OBRA=$id_obra "  ;     // ejecuto DELETE
+       $Conn->query($sql) ;
+
+       $sql= "DELETE FROM `SubObras`  WHERE ID_OBRA=$id_obra AND ID_SUBOBRA<>'$id_subobra_auto' "  ;     // ejecuto DELETE
+       $Conn->query($sql) ;
+    }
 }
 else
 {
